+++ build.gradle
@@ -37,6 +37,8 @@ dependencies {
     testImplementation "io.micronaut:micronaut-inject-groovy"
     testImplementation "io.micronaut.test:micronaut-test-spock"
     testImplementation "io.micronaut.test:micronaut-test-junit5"
+    testCompile "io.micronaut.test:micronaut-test-spock:1.1.3"
+
 }
 
 test.classpath += configurations.developmentOnly
