--- src/main/groovy/gl51/Application.groovy
@@ -1,11 +1,0 @@
-package gl51
-
-import io.micronaut.runtime.Micronaut
-import groovy.transform.CompileStatic
-
-@CompileStatic
-class Application {
-    static void main(String[] args) {
-        Micronaut.run(Application)
-    }
-}
