--- src/main/groovy/gl51/data/Picture.groovy
@@ -1,9 +1,0 @@
-package gl51.data
-
-class Picture {
-    int pixels_x
-    int pixels_y
-    File pictureFile
-    String name
-    String coordonnees
-}
