--- src/main/groovy/gl51/moovie/data/Moovie.groovy
@@ -1,12 +1,0 @@
-package gl51.moovie.data;
-
-import io.micronaut.http.server.netty.binders.$CompletableFutureBodyBinderDefinitionClass;
-
-import java.util.List;
-
-public class Moovie {
-    String imdbID
-    String title
-    List<String> actors
-}
-
