--- src/main/groovy/gl51/moovie/service/Impl/MoovieClientImpl.groovy
@@ -1,4 +1,0 @@
-package gl51.moovie.service.Impl
-
-class MoovieClientImpl implements gl51.moovie.service.MoovieClient {
-}
