--- src/main/groovy/gl51/moovie/service/MoovieClient.groovy
@@ -1,4 +1,0 @@
-package gl51.moovie.service;
-
-public interface MoovieClient {
-}
