+++ src/main/groovy/gl51/movie/data/Movie.groovy
@@ -1,0 +1,12 @@
+package gl51.movie.data;
+
+import io.micronaut.http.server.netty.binders.$CompletableFutureBodyBinderDefinitionClass;
+
+import java.util.List;
+
+public class Movie {
+    String imdbID
+    String title
+    List<String> actors
+}
+
