+++ src/main/groovy/gl51/picture/Application.groovy
@@ -1,0 +1,11 @@
+package gl51.picture
+
+import io.micronaut.runtime.Micronaut
+import groovy.transform.CompileStatic
+
+@CompileStatic
+class Application {
+    static void main(String[] args) {
+        Micronaut.run(Application)
+    }
+}
