+++ src/main/groovy/gl51/picture/data/Picture.groovy
@@ -1,0 +1,9 @@
+package gl51.picture.data
+
+class Picture {
+    int pixels_x
+    int pixels_y
+    File pictureFile
+    String name
+    String coordonnees
+}
