+++ src/main/groovy/gl51/picture/service/Impl/PictureDbUpdateServiceImpl.groovy
@@ -1,0 +1,19 @@
+package gl51.picture.service.Impl
+
+import gl51.picture.data.Picture
+import gl51.picture.service.PictureDbUpdateService
+import gl51.picture.service.PictureService
+
+import javax.inject.Inject
+
+class PictureDbUpdateServiceImpl implements PictureDbUpdateService{
+
+    @Inject PictureService pictureService
+
+    @Override
+    int updateDb(){
+        Picture pic = pictureService.fetchPicture()
+        // DbUpdate(pic)
+        return 0
+    }
+}
