+++ src/main/groovy/gl51/picture/service/Impl/PictureFiligraneServiceImpl.groovy
@@ -1,0 +1,11 @@
+package gl51.picture.service.Impl
+
+import gl51.picture.service.PictureFiligraneService
+
+class PictureFiligraneServiceImpl implements  PictureFiligraneService {
+
+    @Override
+    void apllyFiligrane(File pic){
+        //add filigrane to a picture
+    }
+}
