+++ src/main/groovy/gl51/picture/service/Impl/PictureServiceImpl.groovy
@@ -1,0 +1,12 @@
+package gl51.picture.service.Impl
+
+import gl51.picture.data.Picture
+import gl51.picture.service.PictureService
+
+@Singleton
+
+class PictureServiceImpl implements PictureService {
+    //Récupère l'image
+    @Override
+    Picture fetchPicture(){return null}
+}
