+++ src/main/groovy/gl51/picture/service/PictureCloudTransfertService.groovy
@@ -1,0 +1,6 @@
+package gl51.picture.service
+
+interface PictureCloudTransfertService {
+    //Transfert les images crées vers le cloud
+    int transfert()
+}
