+++ src/main/groovy/gl51/picture/service/PictureDbUpdateService.groovy
@@ -1,0 +1,7 @@
+package gl51.picture.service
+
+interface PictureDbUpdateService {
+    //Met à jour la base de données
+    int updateDb()
+
+}
