+++ src/main/groovy/gl51/picture/service/PictureFiligraneService.groovy
@@ -1,0 +1,6 @@
+package gl51.picture.service
+
+interface PictureFiligraneService {
+    //Applique le Filigrane au thumbnail
+    void apllyFiligrane(File pic)
+}
