+++ src/main/groovy/gl51/picture/service/PictureService.groovy
@@ -1,0 +1,8 @@
+package gl51.picture.service
+
+import gl51.picture.data.Picture
+
+interface PictureService {
+    // Récupère l'image
+    Picture fetchPicture()
+}
