+++ src/main/groovy/gl51/picture/service/PictureThumbnailService.groovy
@@ -1,0 +1,8 @@
+package gl51.picture.service
+
+import gl51.picture.data.Picture
+
+interface PictureThumbnailService {
+    //Crée un Thumbanil de l'image chargée
+    Picture makeThumbnail()
+}
