--- src/main/groovy/gl51/service/Impl/PictureFiligraneServiceImpl.groovy
@@ -1,11 +1,0 @@
-package gl51.service.Impl
-
-import gl51.service.PictureFiligraneService
-
-class PictureFiligraneServiceImpl implements  PictureFiligraneService {
-
-    @Override
-    void apllyFiligrane(File pic){
-        //add filigrane to a picture
-    }
-}
