--- src/main/groovy/gl51/service/Impl/PictureServiceImpl.groovy
@@ -1,12 +1,0 @@
-package gl51.service.Impl
-
-import gl51.data.Picture
-import gl51.service.PictureService
-
-@Singleton
-
-class PictureServiceImpl implements PictureService {
-    //Récupère l'image
-    @Override
-    Picture fetchPicture(){return null}
-}
