--- src/main/groovy/gl51/service/PictureCloudTransfertService.groovy
@@ -1,8 +1,0 @@
-package gl51.service
-
-import gl51.data.Picture
-
-interface PictureCloudTransfertService {
-    //Transfert les images crées vers le cloud
-    int transfert()
-}
