--- src/main/groovy/gl51/service/PictureDbUpdateService.groovy
@@ -1,9 +1,0 @@
-package gl51.service
-
-import gl51.data.Picture
-
-interface PictureDbUpdateService {
-    //Met à jour la base de données
-    int updateDb()
-
-}
