--- src/main/groovy/gl51/service/PictureFiligraneService.groovy
@@ -1,6 +1,0 @@
-package gl51.service
-
-interface PictureFiligraneService {
-    //Applique le Filigrane au thumbnail
-    void apllyFiligrane(File pic)
-}
