--- src/main/groovy/gl51/service/PictureService.groovy
@@ -1,8 +1,0 @@
-package gl51.service
-
-import gl51.data.Picture
-
-interface PictureService {
-    // Récupère l'image
-    Picture fetchPicture()
-}
