--- src/main/groovy/gl51/service/PictureThumbnailService.groovy
@@ -1,8 +1,0 @@
-package gl51.service
-
-import gl51.data.Picture
-
-interface PictureThumbnailService {
-    //Crée un Thumbanil de l'image chargée
-    Picture makeThumbnail()
-}
