--- src/main/groovy/gl51/service/PictureUploadService.groovy
@@ -1,6 +1,0 @@
-package gl51.service
-
-interface PictureUploadService {
-    // Upload les images, les redimensionne, applique le filigrane et met à jour la base de données
-    int uploadPicture()
-}
